Add read module
