+++ swarmcg/optimize_model.py
@@ -1,6 +1,7 @@
 # some numpy version have this ufunc warning at import + many packages call numpy and display annoying warnings
 import warnings
 
+import swarmcg.io.read
 import swarmcg.scoring.distances
 
 warnings.filterwarnings("ignore")
@@ -218,7 +219,7 @@ def run(ns):
     scg.read_ndx_atoms2beads(ns)  # read mapping, get atoms accurences in beads
     scg.get_atoms_weights_in_beads(ns)  # get weights of atoms within beads
 
-    scg.read_cg_itp_file(ns)  # load the ITP object and find out geoms grouping
+    swarmcg.io.read.read_cg_itp_file(ns)  # load the ITP object and find out geoms grouping
     scg.process_scaling_str(ns)  # process the bonds scaling specified by user
 
     print()
